Move `EraGov` interface into `cardano-ledger-core`
