+++ libs/cardano-ledger-core/cardano-ledger-core.cabal
@@ -80,6 +80,7 @@ library
     Cardano.Ledger.Core.TxCert
     Cardano.Ledger.Keys.Internal
     Cardano.Ledger.State.AccountState
+    Cardano.Ledger.State.Governance
     Cardano.Ledger.State.PoolDistr
     Cardano.Ledger.State.SnapShots
     Cardano.Ledger.State.UTxO
