+++ libs/cardano-ledger-core/src/Cardano/Ledger/State.hs
@@ -3,10 +3,12 @@ module Cardano.Ledger.State (
   module Cardano.Ledger.State.UTxO,
   module Cardano.Ledger.State.PoolDistr,
   module Cardano.Ledger.State.SnapShots,
+  module Cardano.Ledger.State.Governance,
 )
 where
 
 import Cardano.Ledger.State.AccountState
+import Cardano.Ledger.State.Governance
 import Cardano.Ledger.State.PoolDistr
 import Cardano.Ledger.State.SnapShots
 import Cardano.Ledger.State.UTxO
